Move new cmc pro methods to separate service.
