+++ app/services/coin_market_cap_service.rb
@@ -1,168 +1,40 @@
 require_relative '../../lib/tasks/batch_process'
 
 class CoinMarketCapService
-  attr_accessor :db_missing_coins
-  attr_accessor :cmc_missing_data
+  attr_accessor :failed_updates
 
   def initialize
-    @db_missing_coins = []
-    @cmc_missing_data = []
+    @failed_updates = []
   end
 
-  def supply_update
-    coins = load_northpole_ticker_data
-    batch_process(coins) do |coin|
+  def ticker_update
+    data = load_cmc_ticker_data
+    batch_process(coins(data)) do |coin|
       identifier = coin['identifier']
-      update_coin_supply(identifier, coin)
+      update_coin(identifier, coin)
     end
-    log_db_missing_coins
-  end
-
-  def ticker_update(start = 1, limit = 100) # 1-indexed
-    coins = load_cmc_latest_data(start, limit)
-    unless coins.nil?
-      batch_process(coins) do |coin|
-        identifier = coin['slug']
-        update_coin_prices(identifier, coin)
-      end
-      log_db_missing_coins
-      log_missing_snap_data
-    end
-  end
-
-  def pairs_update(start = 0, limit = 20) # 0-indexed
-    coins = Coin.top(limit).offset(start)
-    batch_process(coins) do |coin|
-      update_coin_pairs(coin)
-    end
-    log_missing_pairs_data
+    log_failed_updates
   end
 
   private
 
-  def log_db_missing_coins
-    @db_missing_coins.sort! { |left, right| left[:ranking] <=> right[:ranking] }
-    @db_missing_coins.each do |update|
+  def log_failed_updates
+    @failed_updates.sort! { |left, right| left[:ranking] <=> right[:ranking] }
+    @failed_updates.each do |update|
       Rails.logger.info "MISSING COIN: Rank #{update[:ranking]} #{update[:identifier]} coin from CMC is missing from the `coins` table."
     end
-    unless @db_missing_coins.empty?
-      Rollbar.error('Coins table missing CMC coins', :coins => @db_missing_coins)
-    end
   end
 
-  def log_missing_snap_data
-    if @cmc_missing_data.empty?
-      Net::HTTP.get(URI.parse(ENV.fetch('HEALTHCHECK_SNAP_PRICES')))
-    else
-      Net::HTTP.post(URI.parse("#{ENV.fetch('HEALTHCHECK_SNAP_PRICES')}/fail"), @cmc_missing_data.to_json)
-    end
-  end
-
-  def log_missing_pairs_data
-    if @cmc_missing_data.empty?
-      Net::HTTP.get(URI.parse(ENV.fetch('HEALTHCHECK_MARKET_PAIRS')))
-    else
-      Net::HTTP.post(URI.parse("#{ENV.fetch('HEALTHCHECK_MARKET_PAIRS')}/fail"), @cmc_missing_data.to_json)
-    end
-  end
-
-  def has_missing_snap_data(data)
-    quote = data.dig("quote", currency)
-    if quote["price"].blank? || quote["market_cap"].blank? || quote["volume_24h"].blank? ||
-      quote["percent_change_1h"].blank? || quote["percent_change_24h"].blank? ||
-      quote["percent_change_7d"].blank? || data["total_supply"].blank? ||
-      data['circulating_supply'].blank? then
-      return true
-    else
-      return false
-    end
-  end
-
-  def update_coin_prices(identifier, data)
-    if has_missing_snap_data(data)
-      @cmc_missing_data << { identifier: identifier, data: data }
-    else
-      perform_update_prices(data)
-    end
-
+  def update_coin(identifier, data)
     coin = Coin.find_by(slug: identifier)
     if !coin
-      @db_missing_coins << { identifier: identifier, ranking: data['cmc_rank'] }
+      @failed_updates << { identifier: identifier, ranking: data['position'].to_i }
     else
-      perform_update_ranking(coin, data)
+      perform_update(coin, data)
     end
   end
 
-  def update_coin_supply(identifier, data)
-    coin = Coin.find_by(slug: identifier)
-    if !coin
-      @db_missing_coins << { identifier: identifier, ranking: data['position'].to_i }
-    else
-      perform_update_supply(coin, data)
-    end
-  end
-
-  def update_coin_pairs(coin)
-    data = coin.cmc_id ? load_cmc_pair_data(id: coin.cmc_id) : load_cmc_pair_data(symbol: coin.symbol)
-    identifier = coin.slug
-    perform_update_pairs(identifier, data)
-  end
-
-  def perform_update_prices(data)
-    quote = data.dig("quote", currency)
-    coin_hash = {
-      :price => quote["price"],
-      :market_cap => quote["market_cap"],
-      :volume24h => quote["volume_24h"],
-      :change1h => quote["percent_change_1h"],
-      :change24h => quote["percent_change_24h"],
-      :change7d => quote["percent_change_7d"],
-      :total_supply => data["total_supply"],
-      :available_supply => data['circulating_supply'],
-      :max_supply => data['max_supply'],
-      :last_retrieved => Time.now.utc.to_s,
-    }
-    Rails.cache.write("#{data['slug']}:snapshot", coin_hash)
-  end
-
-  def perform_update_pairs(identifier, data)
-    snapshot = Rails.cache.read("#{identifier}:snapshot")
-    base_volume24 = snapshot[:volume24h] unless snapshot.blank?
-
-    raw_market_pairs = data.dig("market_pairs")
-    market_pairs = raw_market_pairs.map do |pair|
-      quote = pair.dig("quote", currency)
-      volume_24h_quote =
-      volume24 = quote["volume_24h"]
-      {
-        :exchange_name => pair.dig("exchange", "name"),
-        :exchange_slug => pair.dig("exchange", "slug"),
-        :pair => pair["market_pair"],
-        :price => quote["price"],
-        :volume24h => volume24,
-        :volume_percentage => (volume24 / base_volume24 unless base_volume24.blank?),
-        :volume24h_quote => pair.dig("quote", "exchange_reported", "volume_24h_quote"),
-        :quote_currency_symbol => pair.dig("market_pair_quote", "currency_symbol"),
-      }
-    end
-
-    coin_hash = {
-      :total_pairs => data["num_market_pairs"],
-      :market_pairs => market_pairs,
-      :last_retrieved => Time.now.utc.to_s,
-    }
-    Rails.cache.write("#{identifier}:pairs", coin_hash)
-  end
-
-  def perform_update_ranking(coin, data)
-    coin.update(
-      ranking: data['cmc_rank'],
-      last_synced: data['last_updated'],
-      ico_status: 'listed'
-    )
-  end
-
-  def perform_update_supply(coin, data)
+  def perform_update(coin, data)
     coin.update(
       ranking: data['position'],
       market_cap: data['marketCap'],
@@ -176,65 +48,13 @@ class CoinMarketCapService
     )
   end
 
-  def load_northpole_ticker_data
+  def coins(data)
+    data['markets']
+  end
+
+  def load_cmc_ticker_data
     ticker_url = 'http://coinmarketcap.northpole.ro/ticker.json?identifier=&version=v8'
     response = HTTParty.get(ticker_url)
-    contents = JSON.parse(response.body)
-
-    contents['markets']
-  end
-
-  def load_cmc_latest_data(start, limit)
-    ticker_url = "https://pro-api.coinmarketcap.com/v1/cryptocurrency/listings/latest"
-    query = { :start => start, :limit => limit }
-    headers = { "X-CMC_PRO_API_KEY" => ENV.fetch('COINMARKETCAP_API_KEY') }
-    response = HTTParty.get(ticker_url, :query => query, :headers => headers)
-    contents = JSON.parse(response.body)
-
-    # ping health check if api error
-    json_response_code = get_json_response_code(contents)
-
-    if response.success? && json_response_code == 0 then
-      Net::HTTP.get(URI.parse(ENV.fetch('HEALTHCHECK_SNAP_PRICES')))
-      return contents['data']
-    else
-      error_message = get_error_message(contents)
-      Net::HTTP.post(URI.parse("#{ENV.fetch('HEALTHCHECK_SNAP_PRICES')}/fail"), "ERROR HTTP(#{response.code}) JSON(#{json_response_code}): #{error_message}")
-      return nil
-    end
-  end
-
-  def load_cmc_pair_data(id: nil, symbol: nil)
-    return nil if id.nil? and symbol.nil?
-
-    api_url = "https://pro-api.coinmarketcap.com/v1/cryptocurrency/market-pairs/latest"
-    query = if id.nil? then { :symbol => symbol } else { :id => id } end
-    headers = { "X-CMC_PRO_API_KEY" => ENV.fetch('COINMARKETCAP_API_KEY') }
-    response = HTTParty.get(api_url, :query => query, :headers => headers)
-    contents = JSON.parse(response.body)
-
-    # ping health check if api error
-    json_response_code = get_json_response_code(contents)
-
-    if response.success? && json_response_code == 0 then
-      Net::HTTP.get(URI.parse(ENV.fetch('HEALTHCHECK_MARKET_PAIRS')))
-      return contents['data']
-    else
-      error_message = get_error_message(contents)
-      Net::HTTP.post(URI.parse("#{ENV.fetch('HEALTHCHECK_MARKET_PAIRS')}/fail"), "ERROR HTTP(#{response.code}) JSON(#{json_response_code}): #{error_message}")
-      return nil
-    end
-  end
-
-  def get_json_response_code(contents)
-    contents.dig('status', 'error_code') || 0
-  end
-
-  def get_error_message(contents)
-    contents.dig('status', 'error_message')
-  end
-
-  def currency
-    return "USD"
+    JSON.parse(response.body)
   end
 end
